fixup! [libc++] Use correct size for deallocation of arrays in shared_ptr
use int instead of int64_t to avoid the need for an include

+++ libcxx/test/libcxx/memory/shared_ptr_array.pass.cpp
@@ -17,11 +17,11 @@
 #include <memory>
 
 int main(int, char**) {
-  std::allocate_shared<int64_t[]>(std::allocator<int64_t>{}, 10);
-  std::make_shared<int64_t[]>(10);
+  std::allocate_shared<int[]>(std::allocator<int>{}, 10);
+  std::make_shared<int[]>(10);
 
-  std::allocate_shared<int64_t[10]>(std::allocator<int64_t>{});
-  std::make_shared<int64_t[10]>();
+  std::allocate_shared<int[10]>(std::allocator<int>{});
+  std::make_shared<int[10]>();
 
   return 0;
 }
